+++ ksp/sealed/src/commonMain/kotlin/OldGenerateSealedWorkaround.kt
@@ -1,0 +1,6 @@
+package dev.inmo.microutils.kps.sealed
+
+import dev.inmo.micro_utils.ksp.sealed.GenerateSealedWorkaround
+
+@Deprecated("Replaced", ReplaceWith("GenerateSealedWorkaround", "dev.inmo.micro_utils.ksp.sealed.GenerateSealedWorkaround"))
+typealias GenerateSealedWorkaround = GenerateSealedWorkaround